(Refactoring GOD Class) Extract class from GameManager to BVManager.java
